data-generator/requirements.txt
kafka-python==2.0.2
Faker==18.9.0
python-json-logger==2.0.7